Initial minimal end-to-end schema processing with associated TODOs
