+++ adapters/handlers/rest/state/state.go
@@ -35,6 +35,8 @@ import (
 	"github.com/weaviate/weaviate/usecases/schema"
 	"github.com/weaviate/weaviate/usecases/sharding"
 	"github.com/weaviate/weaviate/usecases/traverser"
+
+	schemav2 "github.com/weaviate/weaviate/cloud/store"
 )
 
 // State is the only source of application-wide state
@@ -65,6 +67,10 @@ type State struct {
 	BatchManager       *objects.BatchManager
 	ClusterHttpClient  *http.Client
 	ReindexCtxCancel   context.CancelFunc
+	/// TODO-RAFT START
+	MetaStore   *schemav2.Store
+	MetaHandler *schema.Handler
+	/// TODO-RAFT END
 }
 
 // GetGraphQL is the safe way to retrieve GraphQL from the state as it can be
