+++ usecases/schema/class_test.go
@@ -1,0 +1,12 @@
+package schema
+
+/// TODO-RAFT START
+// Fix Unit Tests
+// With class-related transactions refactored into class.go, we need to re-implement the following test cases:
+// - Get schema (previously located in get_test.go)
+// - Validation (previously located in validation_test.go)
+// - Adding new classes (previously in add_test.go)
+// - Updating existing classes (previously in update_test.go)
+// - Restore classes once Restore is implemented (previously in restore_test.go)
+// Please consult the previous implementation's test files for reference.
+/// TODO-RAFT END
