--- usecases/schema/delete_property.go
@@ -1,33 +1,0 @@
-//                           _       _
-// __      _____  __ ___   ___  __ _| |_ ___
-// \ \ /\ / / _ \/ _` \ \ / / |/ _` | __/ _ \
-//  \ V  V /  __/ (_| |\ V /| | (_| | ||  __/
-//   \_/\_/ \___|\__,_| \_/ |_|\__,_|\__\___|
-//
-//  Copyright © 2016 - 2023 Weaviate B.V. All rights reserved.
-//
-//  CONTACT: hello@weaviate.io
-//
-
-package schema
-
-import (
-	"context"
-	"fmt"
-
-	"github.com/weaviate/weaviate/entities/models"
-)
-
-// DeleteClassProperty from existing Schema
-func (m *Manager) DeleteClassProperty(ctx context.Context, principal *models.Principal,
-	class string, property string,
-) error {
-	err := m.Authorizer.Authorize(principal, "update", "schema/objects")
-	if err != nil {
-		return err
-	}
-
-	return fmt.Errorf("deleting a property is currently not supported, see " +
-		"https://github.com/weaviate/weaviate/issues/973 for details.")
-	// return m.deleteClassProperty(ctx, class, property, kind.Action)
-}
