+++ usecases/schema/property_test.go
@@ -1,0 +1,10 @@
+package schema
+
+/// TODO-RAFT START
+// Fix Unit Tests
+// With property-related transactions refactored into property.go, we need to re-implement the following test cases:
+// - Adding new property (previously in add_property_test.go)
+// - Updating existing property 
+// - Validation (previously in remove_duplicate_props_test.go)
+// Please consult the previous implementation's test files for reference.
+/// TODO-RAFT END
